LPS-54271 move to monitor.internal instead of monitor.internal.statistics
