+++ modules/portal/portal-monitoring/src/com/liferay/portal/monitoring/internal/DataSampleFactoryImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.monitoring.internal.statistics;
+package com.liferay.portal.monitoring.internal;
 
 import com.liferay.portal.kernel.monitoring.DataSample;
 import com.liferay.portal.kernel.monitoring.DataSampleFactory;
