+++ modules/portal/portal-monitoring/src/com/liferay/portal/monitoring/internal/statistics/service/ServiceRequestDataSample.java
@@ -16,7 +16,7 @@ package com.liferay.portal.monitoring.internal.statistics.service;
 
 import com.liferay.portal.kernel.monitoring.MethodSignature;
 import com.liferay.portal.monitoring.MonitorNames;
-import com.liferay.portal.monitoring.internal.statistics.BaseDataSample;
+import com.liferay.portal.monitoring.internal.BaseDataSample;
 
 /**
  * @author Michael C. Han
